requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
playwright>=1.35.0
redis>=5.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
tiktoken>=0.5.0
httpx
